Refactored aamped and maamped tests
