--- tests/test_aamped_two_constant_subsequences.py
@@ -1,61 +1,0 @@
-import numpy as np
-import numpy.testing as npt
-import pandas as pd
-from stumpy import core, aamped
-from dask.distributed import Client, LocalCluster
-import pytest
-import warnings
-import naive
-
-
-@pytest.fixture(scope="module")
-def dask_cluster():
-    cluster = LocalCluster(n_workers=2, threads_per_worker=2)
-    yield cluster
-    cluster.close()
-
-
-@pytest.mark.filterwarnings("ignore:\\s+A large number of values are smaller")
-@pytest.mark.filterwarnings("ignore:\\s+For a self-join")
-@pytest.mark.filterwarnings("ignore:numpy.dtype size changed")
-@pytest.mark.filterwarnings("ignore:numpy.ufunc size changed")
-@pytest.mark.filterwarnings("ignore:numpy.ndarray size changed")
-@pytest.mark.filterwarnings("ignore:\\s+Port 8787 is already in use:UserWarning")
-def test_two_constant_subsequences_A_B_join(dask_cluster):
-    with Client(dask_cluster) as dask_client:
-        T_A = np.concatenate(
-            (np.zeros(10, dtype=np.float64), np.ones(10, dtype=np.float64))
-        )
-        T_B = np.concatenate(
-            (np.zeros(20, dtype=np.float64), np.ones(5, dtype=np.float64))
-        )
-        m = 3
-        ref_mp = naive.aamp(T_A, m, T_B=T_B)
-        comp_mp = aamped(dask_client, T_A, m, T_B, ignore_trivial=False)
-        naive.replace_inf(ref_mp)
-        naive.replace_inf(comp_mp)
-        npt.assert_almost_equal(ref_mp[:, 0], comp_mp[:, 0])  # ignore indices
-
-
-@pytest.mark.filterwarnings("ignore:\\s+A large number of values are smaller")
-@pytest.mark.filterwarnings("ignore:\\s+For a self-join")
-@pytest.mark.filterwarnings("ignore:numpy.dtype size changed")
-@pytest.mark.filterwarnings("ignore:numpy.ufunc size changed")
-@pytest.mark.filterwarnings("ignore:numpy.ndarray size changed")
-@pytest.mark.filterwarnings("ignore:\\s+Port 8787 is already in use:UserWarning")
-def test_two_constant_subsequences_A_B_join_df(dask_cluster):
-    with Client(dask_cluster) as dask_client:
-        T_A = np.concatenate(
-            (np.zeros(10, dtype=np.float64), np.ones(10, dtype=np.float64))
-        )
-        T_B = np.concatenate(
-            (np.zeros(20, dtype=np.float64), np.ones(5, dtype=np.float64))
-        )
-        m = 3
-        ref_mp = naive.aamp(T_A, m, T_B=T_B)
-        comp_mp = aamped(
-            dask_client, pd.Series(T_A), m, pd.Series(T_B), ignore_trivial=False
-        )
-        naive.replace_inf(ref_mp)
-        naive.replace_inf(comp_mp)
-        npt.assert_almost_equal(ref_mp[:, 0], comp_mp[:, 0])  # ignore indices
